refactor relative time logic to use new utility function
